# Conflicts:
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/DailyRate.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/Data.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/Detail.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/FetchServicesResponsePojo.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/Price.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/ResponseData.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/ServiceDetail.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/ServiceOption.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/ServiceOptionDesc.java
#	app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/ValidationError.java

+++ app/src/main/java/com/ehg/booking/hotel/pojo/fetchservicesresponsepojo/DailyRate.java
@@ -1,29 +1,50 @@
+/*
+ *  Created by Emaar Hospitality Group on 17/10/18 6:00 PM
+ *  Copyright (C) 2018  All rights reserved.
+ *  Last modified 17/10/18 6:00 PM
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
 
 package com.ehg.booking.hotel.pojo.fetchservicesresponsepojo;
 
 import com.google.gson.annotations.SerializedName;
 import java.util.List;
 
+/**
+ * DailyRate.
+ */
 public class DailyRate {
 
   @SerializedName("AdultQuantity")
-  private int adultQuantity;
+  private float adultQuantity;
   @SerializedName("AdultRate")
-  private int adultRate;
+  private float adultRate;
   @SerializedName("AdultUnitPrice")
-  private int adultUnitPrice;
+  private float adultUnitPrice;
   @SerializedName("ChildQuantity")
-  private int childQuantity;
+  private float childQuantity;
   @SerializedName("ChildRate")
-  private int childRate;
+  private float childRate;
   @SerializedName("ChildUnitPrice")
-  private int childUnitPrice;
+  private float childUnitPrice;
   @SerializedName("DiscountRate")
-  private int discountRate;
+  private float discountRate;
   @SerializedName("Price")
   private List<Price> price;
   @SerializedName("Quantity")
-  private int quantity;
+  private float quantity;
   @SerializedName("ServiceDate")
   private String serviceDate;
   @SerializedName("ServiceOptionRefId")
@@ -36,14 +57,14 @@ public class DailyRate {
    *
    * @return Gets the value of adultQuantity and returns adultQuantity.
    */
-  public int getAdultQuantity() {
+  public float getAdultQuantity() {
     return adultQuantity;
   }
 
   /**
    * Sets the adultQuantity. You can use getAdultQuantity() to get the value of adultQuantity.
    */
-  public void setAdultQuantity(int adultQuantity) {
+  public void setAdultQuantity(float adultQuantity) {
     this.adultQuantity = adultQuantity;
   }
 
@@ -52,14 +73,14 @@ public class DailyRate {
    *
    * @return Gets the value of adultRate and returns adultRate.
    */
-  public int getAdultRate() {
+  public float getAdultRate() {
     return adultRate;
   }
 
   /**
    * Sets the adultRate. You can use getAdultRate() to get the value of adultRate.
    */
-  public void setAdultRate(int adultRate) {
+  public void setAdultRate(float adultRate) {
     this.adultRate = adultRate;
   }
 
@@ -68,14 +89,14 @@ public class DailyRate {
    *
    * @return Gets the value of adultUnitPrice and returns adultUnitPrice.
    */
-  public int getAdultUnitPrice() {
+  public float getAdultUnitPrice() {
     return adultUnitPrice;
   }
 
   /**
    * Sets the adultUnitPrice. You can use getAdultUnitPrice() to get the value of adultUnitPrice.
    */
-  public void setAdultUnitPrice(int adultUnitPrice) {
+  public void setAdultUnitPrice(float adultUnitPrice) {
     this.adultUnitPrice = adultUnitPrice;
   }
 
@@ -84,14 +105,14 @@ public class DailyRate {
    *
    * @return Gets the value of childQuantity and returns childQuantity.
    */
-  public int getChildQuantity() {
+  public float getChildQuantity() {
     return childQuantity;
   }
 
   /**
    * Sets the childQuantity. You can use getChildQuantity() to get the value of childQuantity.
    */
-  public void setChildQuantity(int childQuantity) {
+  public void setChildQuantity(float childQuantity) {
     this.childQuantity = childQuantity;
   }
 
@@ -100,14 +121,14 @@ public class DailyRate {
    *
    * @return Gets the value of childRate and returns childRate.
    */
-  public int getChildRate() {
+  public float getChildRate() {
     return childRate;
   }
 
   /**
    * Sets the childRate. You can use getChildRate() to get the value of childRate.
    */
-  public void setChildRate(int childRate) {
+  public void setChildRate(float childRate) {
     this.childRate = childRate;
   }
 
@@ -116,14 +137,14 @@ public class DailyRate {
    *
    * @return Gets the value of childUnitPrice and returns childUnitPrice.
    */
-  public int getChildUnitPrice() {
+  public float getChildUnitPrice() {
     return childUnitPrice;
   }
 
   /**
    * Sets the childUnitPrice. You can use getChildUnitPrice() to get the value of childUnitPrice.
    */
-  public void setChildUnitPrice(int childUnitPrice) {
+  public void setChildUnitPrice(float childUnitPrice) {
     this.childUnitPrice = childUnitPrice;
   }
 
@@ -132,14 +153,14 @@ public class DailyRate {
    *
    * @return Gets the value of discountRate and returns discountRate.
    */
-  public int getDiscountRate() {
+  public float getDiscountRate() {
     return discountRate;
   }
 
   /**
    * Sets the discountRate. You can use getDiscountRate() to get the value of discountRate.
    */
-  public void setDiscountRate(int discountRate) {
+  public void setDiscountRate(float discountRate) {
     this.discountRate = discountRate;
   }
 
@@ -155,7 +176,8 @@ public class DailyRate {
   /**
    * Sets the price. You can use getPrice() to get the value of price.
    */
-  public void setPrice(List<Price> price) {
+  public void setPrice(
+      List<Price> price) {
     this.price = price;
   }
 
@@ -164,14 +186,14 @@ public class DailyRate {
    *
    * @return Gets the value of quantity and returns quantity.
    */
-  public int getQuantity() {
+  public float getQuantity() {
     return quantity;
   }
 
   /**
    * Sets the quantity. You can use getQuantity() to get the value of quantity.
    */
-  public void setQuantity(int quantity) {
+  public void setQuantity(float quantity) {
     this.quantity = quantity;
   }
 
@@ -201,7 +223,8 @@ public class DailyRate {
   }
 
   /**
-   * Sets the serviceOptionRefId. You can use getServiceOptionRefId() to get the value of serviceOptionRefId.
+   * Sets the serviceOptionRefId. You can use getServiceOptionRefId() to get the value of
+   * serviceOptionRefId.
    */
   public void setServiceOptionRefId(String serviceOptionRefId) {
     this.serviceOptionRefId = serviceOptionRefId;
